Update for lighttpd and "DB" class not found
Update the information that "lighttpd" server can be used.
Added the information about the PHP "DB" class not found error in
reason that meekrodb is integrated as a submodule and has to be
checkout using `git clone --recurse-submodules https://github.com/jrie/sqstorage`
or as described here `https://github.com/jrie/sqstorage/issues/106`.